+++ changes.rst
@@ -3,7 +3,8 @@ Changes
 2016-09-05
 ----------
 
-- Python 3 support not yet working properly
+- release v0.2.4
+- roll back Python 3 support as not yet working properly
 - add translate methods to IDF
 - add getsurfaces by surface type method to IDF